refactor: extract bright header to partial
