+++ src/components/Tooltip/Tooltip.tsx
@@ -1,33 +1,13 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
 import React, { cloneElement, FC, ReactElement, ReactNode } from "react";
-import { merge } from "@utilities/merge";
 import { Button, ButtonStyle, ButtonSize } from "@elements/Button/Button";
 import { IconSize } from "@elements/Icon/IconSize";
-import IconInfo from "@elements/Icon/Generated/IconInfo";
-import IconCheck from "@elements/Icon/Generated/IconCheck";
-import IconCallout from "@elements/Icon/Generated/IconCallout";
-import IconDocument from "@elements/Icon/Generated/IconDocument";
+import { BrightHeader, BrightHeaderVariants } from "./partials/BrightHeader";
 
-export enum BrightHeaderVariants {
-    Information = "information",
-    Warning = "warning",
-    Tip = "tip",
-    Note = "note",
-}
-
-const brightHeaderBackgroundColors: Record<BrightHeaderVariants, string> = {
-    [BrightHeaderVariants.Information]: "tw-bg-violet-60",
-    [BrightHeaderVariants.Warning]: "tw-bg-red-60",
-    [BrightHeaderVariants.Tip]: "tw-bg-green-60",
-    [BrightHeaderVariants.Note]: "tw-bg-yellow-60",
-};
-
-const brightHeaderIcon: Record<BrightHeaderVariants, ReactElement> = {
-    [BrightHeaderVariants.Information]: <IconInfo size={IconSize.Size20} />,
-    [BrightHeaderVariants.Warning]: <IconCallout size={IconSize.Size20} />,
-    [BrightHeaderVariants.Tip]: <IconCheck size={IconSize.Size20} />,
-    [BrightHeaderVariants.Note]: <IconDocument size={IconSize.Size20} />,
+export type TooltipButton = {
+    label: string;
+    action: () => void;
 };
 
 export type TooltipProps = {
@@ -38,27 +18,13 @@ export type TooltipProps = {
     buttons?: [TooltipButton, TooltipButton];
 };
 
-export type TooltipButton = {
-    label: string;
-    action: () => void;
-};
-
 export const Tooltip: FC<TooltipProps> = ({ tooltip, brightHeader, buttons, heading, headingIcon }) => {
     const showButtons = buttons !== undefined && Object.keys(buttons).length === 2;
 
     return (
         <>
             <div className="tw-inline-block tw-max-w-[200px] tw-bg-black-100 dark:tw-bg-white tw-rounded-md tw-shadow-mid tw-text-white dark:tw-text-black-100 tw-z-20">
-                {brightHeader && (
-                    <div
-                        className={merge([
-                            "tw-h-9 tw-flex tw-justify-center tw-items-center tw-rounded-t-md tw-text-white",
-                            brightHeaderBackgroundColors[brightHeader],
-                        ])}
-                    >
-                        {brightHeaderIcon[brightHeader]}
-                    </div>
-                )}
+                {brightHeader && <BrightHeader variant={brightHeader} />}
                 <div data-test-id="tooltip" className="tw-p-4">
                     {heading && headingIcon && (
                         <h4 className="tw-flex tw-text-m tw-font-bold tw-mb-1">
